Removed Analysis Utils
