+++ src/main/resources/application.conf
@@ -74,24 +74,24 @@ akka {
 			receive = off
 		}
 		serialization-bindings {
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestBuilderId$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$ClusterStatusRequest$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestPartitionId$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestSpoutId$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestAnalysisId$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$PartitionsCount" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$ClusterStatusResponse" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$AssignedId" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$BuilderUp" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$SpoutUp" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$AnalysisManagerUp" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$PartitionUp" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestPartitionCount$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestBuilderId$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$ClusterStatusRequest$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestPartitionId$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestSpoutId$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestAnalysisId$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$PartitionsCount" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$ClusterStatusResponse" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$AssignedId" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$BuilderUp" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$SpoutUp" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$AnalysisManagerUp" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$PartitionUp" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestPartitionCount$" = kryo
 
-			"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$ProbeWatermark$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$WatermarkTime$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$SaveState$" = kryo
-			"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$WatermarkTime" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$ProbeWatermark$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$WatermarkTime$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$SaveState$" = kryo
+			"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$WatermarkTime" = kryo
 
 			"com.raphtory.core.actors.spout.SpoutAgent$CommonMessage$SpoutOnline$" = kryo
 			"com.raphtory.core.actors.spout.SpoutAgent$CommonMessage$WorkPlease$" = kryo
@@ -413,25 +413,25 @@ akka-kryo-serialization {
 	# This section is ignored   for idstartegy=default
 	# This section is optional  for idstartegy=explicit
 	classes = [
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestBuilderId$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$ClusterStatusRequest$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestPartitionId$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestSpoutId$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestAnalysisId$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$PartitionsCount"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$ClusterStatusResponse"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$AssignedId"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$BuilderUp"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$SpoutUp"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$AnalysisManagerUp"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$PartitionUp"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatchDog$Message$RequestPartitionCount$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestBuilderId$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$ClusterStatusRequest$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestPartitionId$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestSpoutId$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestAnalysisId$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$PartitionsCount"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$ClusterStatusResponse"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$AssignedId"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$BuilderUp"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$SpoutUp"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$AnalysisManagerUp"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$PartitionUp"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog$Message$RequestPartitionCount$"
 
 
-		"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$ProbeWatermark$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$WatermarkTime$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$SaveState$"
-		"com.raphtory.core.actors.orchestration.clustermanager.WatermarkManager$Message$WatermarkTime"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$ProbeWatermark$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$WatermarkTime$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$SaveState$"
+		"com.raphtory.core.actors.orchestration.raphtoryleader.WatermarkManager$Message$WatermarkTime"
 
 
 		"com.raphtory.core.actors.spout.SpoutAgent$CommonMessage$SpoutOnline$"
