+++ src/main/scala/com/raphtory/core/actors/analysismanager/tasks/RangeAnalysisTask.scala
@@ -1,0 +1,9 @@
+package com.raphtory.core.actors.analysismanager.tasks
+
+import com.raphtory.core.analysis.api.{AggregateSerialiser, Analyser}
+
+final case class RangeAnalysisTask(jobID: String, args: Array[String], analyser: Analyser[Any], serialiser: AggregateSerialiser, start: Long, end: Long, jump: Long, windows: List[Long])
+  extends AnalysisTask(jobID: String, args, analyser, serialiser) {
+  override protected def buildSubTaskController(readyTimestamp: Long): SubTaskController =
+    SubTaskController.fromRangeTask(start, end, jump, windows)
+}
