+++ src/main/scala/com/raphtory/core/actors/orchestration/componentconnector/SpoutConnector.scala
@@ -4,7 +4,7 @@ import akka.actor.Props
 import akka.cluster.pubsub.DistributedPubSubMediator
 import com.raphtory.core.actors.spout.{Spout, SpoutAgent}
 import akka.pattern.ask
-import com.raphtory.core.actors.orchestration.clustermanager.WatchDog.Message.RequestSpoutId
+import com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog.Message.RequestSpoutId
 
 import scala.concurrent.Future
