+++ src/main/scala/com/raphtory/core/actors/orchestration/raphtoryleader/SeedActor.scala
@@ -1,4 +1,4 @@
-package com.raphtory.core.actors.orchestration.clustermanager
+package com.raphtory.core.actors.orchestration.raphtoryleader
 
 import akka.cluster.Cluster
 import akka.cluster.ClusterEvent._
