+++ src/main/scala/com/raphtory/core/actors/partitionmanager/PartitionManager.scala
@@ -7,7 +7,7 @@ import akka.cluster.pubsub.{DistributedPubSub, DistributedPubSubMediator}
 import com.raphtory.core.actors.RaphtoryActor
 import com.raphtory.core.actors.RaphtoryActor.partitionsPerServer
 import com.raphtory.core.actors.analysismanager.tasks.AnalysisTask.Message.{ReaderWorkersAck, ReaderWorkersOnline}
-import com.raphtory.core.actors.orchestration.clustermanager.WatchDog.Message.PartitionUp
+import com.raphtory.core.actors.orchestration.raphtoryleader.WatchDog.Message.PartitionUp
 import com.raphtory.core.model.communication._
 import com.raphtory.core.model.graph.GraphPartition
